requirements.txt
# Para generar, pip freeze > requirements.txt
# Para instalar, pip install -r requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
django-cors-headers==3.13.0
